lint: split base.go into more files
