+++ lint/base.go
@@ -21,12 +21,6 @@ import (
 	"github.com/zmap/zlint/util"
 )
 
-var (
-	// Lints is a map of all known lints by name. Add a Lint to the map by calling
-	// RegisterLint.
-	Lints = make(map[string]*Lint)
-)
-
 // LintInterface is implemented by each Lint.
 type LintInterface interface {
 	// Initialize runs once per-lint. It is called during RegisterLint().
@@ -41,57 +35,6 @@ type LintInterface interface {
 	// Execute() is the body of the lint. It is called for every certificate for
 	// which CheckApplies() returns true.
 	Execute(c *x509.Certificate) *LintResult
-}
-
-// An Enum to programmatically represent the source of a lint
-type LintSource int
-
-// NOTE(@cpu): If you are adding a new LintSource make sure you have considered
-// updating the Directory() function.
-const (
-	UnknownLintSource LintSource = iota
-	CABFBaselineRequirements
-	RFC5280
-	RFC5480
-	RFC5891
-	ZLint
-	AWSLabs
-	EtsiEsi // ETSI - Electronic Signatures and Infrastructures (ESI)
-	CABFEVGuidelines
-	AppleCTPolicy          // https://support.apple.com/en-us/HT205280
-	MozillaRootStorePolicy // https://github.com/mozilla/pkipolicy
-)
-
-// LintSources contains a list of the valid lint sources we expect to be used
-// by ZLint lints.
-var LintSources = []LintSource{
-	CABFBaselineRequirements,
-	CABFEVGuidelines,
-	RFC5280,
-	RFC5480,
-	RFC5891,
-	AppleCTPolicy,
-	EtsiEsi,
-	ZLint,
-	AWSLabs,
-}
-
-// Directory returns the directory name in `lints/` for the LintSource.
-func (l LintSource) Directory() string {
-	switch l {
-	case CABFBaselineRequirements:
-		return "cabf_br"
-	case CABFEVGuidelines:
-		return "cabf_ev"
-	case RFC5280, RFC5480, RFC5891:
-		return "rfc"
-	case AppleCTPolicy:
-		return "apple"
-	case EtsiEsi:
-		return "etsi"
-	default:
-		return "community"
-	}
 }
 
 // A Lint struct represents a single lint, e.g.
@@ -150,13 +93,4 @@ func (l *Lint) Execute(cert *x509.Certificate) *LintResult {
 	}
 	res := l.Lint.Execute(cert)
 	return res
-}
-
-// RegisterLint must be called once for each lint to be excuted. Duplicate lint
-// names are squashed. Normally, RegisterLint is called during init().
-func RegisterLint(l *Lint) {
-	if err := l.Lint.Initialize(); err != nil {
-		panic("could not initialize lint: " + l.Name + ": " + err.Error())
-	}
-	Lints[l.Name] = l
 }
